// src/lib/api/models/TemplateCreate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * LUCO MINTOS
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TemplateCreate
 */
export interface TemplateCreate {
    /**
     * 
     * @type {string}
     * @memberof TemplateCreate
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof TemplateCreate
     */
    content: string;
    /**
     * 
     * @type {boolean}
     * @memberof TemplateCreate
     */
    _default?: boolean;
    /**
     * 
     * @type {string}
     * @memberof TemplateCreate
     */
    tag?: string;
}

/**
 * Check if a given object implements the TemplateCreate interface.
 */
export function instanceOfTemplateCreate(value: object): value is TemplateCreate {
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('content' in value) || value['content'] === undefined) return false;
    return true;
}

export function TemplateCreateFromJSON(json: any): TemplateCreate {
    return TemplateCreateFromJSONTyped(json, false);
}

export function TemplateCreateFromJSONTyped(json: any, ignoreDiscriminator: boolean): TemplateCreate {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'],
        'content': json['content'],
        '_default': json['default'] == null ? undefined : json['default'],
        'tag': json['tag'] == null ? undefined : json['tag'],
    };
}

export function TemplateCreateToJSON(json: any): TemplateCreate {
    return TemplateCreateToJSONTyped(json, false);
}

export function TemplateCreateToJSONTyped(value?: TemplateCreate | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'content': value['content'],
        'default': value['_default'],
        'tag': value['tag'],
    };
}