// src/lib/api/models/TemplateUpdate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * LUCO MINTOS
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TemplateUpdate
 */
export interface TemplateUpdate {
    /**
     * 
     * @type {string}
     * @memberof TemplateUpdate
     */
    name?: string | null;
    /**
     * 
     * @type {string}
     * @memberof TemplateUpdate
     */
    content?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof TemplateUpdate
     */
    _default?: boolean | null;
    /**
     * 
     * @type {string}
     * @memberof TemplateUpdate
     */
    tag?: string | null;
}

/**
 * Check if a given object implements the TemplateUpdate interface.
 */
export function instanceOfTemplateUpdate(value: object): value is TemplateUpdate {
    return true;
}

export function TemplateUpdateFromJSON(json: any): TemplateUpdate {
    return TemplateUpdateFromJSONTyped(json, false);
}

export function TemplateUpdateFromJSONTyped(json: any, ignoreDiscriminator: boolean): TemplateUpdate {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'] == null ? undefined : json['name'],
        'content': json['content'] == null ? undefined : json['content'],
        '_default': json['default'] == null ? undefined : json['default'],
        'tag': json['tag'] == null ? undefined : json['tag'],
    };
}

export function TemplateUpdateToJSON(json: any): TemplateUpdate {
    return TemplateUpdateToJSONTyped(json, false);
}

export function TemplateUpdateToJSONTyped(value?: TemplateUpdate | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'content': value['content'],
        'default': value['_default'],
        'tag': value['tag'],
    };
}