// src/lib/api/models/TemplatesPublic.ts
/* tslint:disable */
/* eslint-disable */
/**
 * LUCO MINTOS
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { TemplatePublic } from './TemplatePublic';
import {
    TemplatePublicFromJSON,
    TemplatePublicFromJSONTyped,
    TemplatePublicToJSON,
    TemplatePublicToJSONTyped,
} from './TemplatePublic';

/**
 * 
 * @export
 * @interface TemplatesPublic
 */
export interface TemplatesPublic {
    /**
     * 
     * @type {Array<TemplatePublic>}
     * @memberof TemplatesPublic
     */
    data: Array<TemplatePublic>;
    /**
     * 
     * @type {number}
     * @memberof TemplatesPublic
     */
    count: number;
}

/**
 * Check if a given object implements the TemplatesPublic interface.
 */
export function instanceOfTemplatesPublic(value: object): value is TemplatesPublic {
    if (!('data' in value) || value['data'] === undefined) return false;
    if (!('count' in value) || value['count'] === undefined) return false;
    return true;
}

export function TemplatesPublicFromJSON(json: any): TemplatesPublic {
    return TemplatesPublicFromJSONTyped(json, false);
}

export function TemplatesPublicFromJSONTyped(json: any, ignoreDiscriminator: boolean): TemplatesPublic {
    if (json == null) {
        return json;
    }
    return {
        
        'data': ((json['data'] as Array<any>).map(TemplatePublicFromJSON)),
        'count': json['count'],
    };
}

export function TemplatesPublicToJSON(json: any): TemplatesPublic {
    return TemplatesPublicToJSONTyped(json, false);
}

export function TemplatesPublicToJSONTyped(value?: TemplatesPublic | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'data': ((value['data'] as Array<any>).map(TemplatePublicToJSON)),
        'count': value['count'],
    };
}