// src/lib/api/models/TemplatePublic.ts
/* tslint:disable */
/* eslint-disable */
/**
 * LUCO MINTOS
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TemplatePublic
 */
export interface TemplatePublic {
    /**
     * 
     * @type {string}
     * @memberof TemplatePublic
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof TemplatePublic
     */
    content: string;
    /**
     * 
     * @type {boolean}
     * @memberof TemplatePublic
     */
    _default?: boolean;
    /**
     * 
     * @type {string}
     * @memberof TemplatePublic
     */
    tag?: string;
    /**
     * 
     * @type {string}
     * @memberof TemplatePublic
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof TemplatePublic
     */
    ownerId: string;
    /**
     * 
     * @type {Date}
     * @memberof TemplatePublic
     */
    createdAt: Date;
}

/**
 * Check if a given object implements the TemplatePublic interface.
 */
export function instanceOfTemplatePublic(value: object): value is TemplatePublic {
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('content' in value) || value['content'] === undefined) return false;
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('ownerId' in value) || value['ownerId'] === undefined) return false;
    if (!('createdAt' in value) || value['createdAt'] === undefined) return false;
    return true;
}

export function TemplatePublicFromJSON(json: any): TemplatePublic {
    return TemplatePublicFromJSONTyped(json, false);
}

export function TemplatePublicFromJSONTyped(json: any, ignoreDiscriminator: boolean): TemplatePublic {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'],
        'content': json['content'],
        '_default': json['default'] == null ? undefined : json['default'],
        'tag': json['tag'] == null ? undefined : json['tag'],
        'id': json['id'],
        'ownerId': json['owner_id'],
        'createdAt': (new Date(json['created_at'])),
    };
}

export function TemplatePublicToJSON(json: any): TemplatePublic {
    return TemplatePublicToJSONTyped(json, false);
}

export function TemplatePublicToJSONTyped(value?: TemplatePublic | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'content': value['content'],
        'default': value['_default'],
        'tag': value['tag'],
        'id': value['id'],
        'owner_id': value['ownerId'],
        'created_at': value['createdAt'].toISOString(),
    };
}